refactor: move all inflation logic into server/api/handler folder
